# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VitalFisio Clinic Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface LoginResponse {
  id: number;
  username: string;
  name: string;
  role: string;
}

export type PatientInsuranceType =
  (typeof PatientInsuranceType)[keyof typeof PatientInsuranceType];

export const PatientInsuranceType = {
  particular: "particular",
  convenio: "convenio",
} as const;

export interface Patient {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  birthDate?: string | null;
  insuranceType: PatientInsuranceType;
  /** @nullable */
  insuranceName?: string | null;
  totalSessions: number;
  remainingSessions: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreatePatientBodyInsuranceType =
  (typeof CreatePatientBodyInsuranceType)[keyof typeof CreatePatientBodyInsuranceType];

export const CreatePatientBodyInsuranceType = {
  particular: "particular",
  convenio: "convenio",
} as const;

export interface CreatePatientBody {
  name: string;
  phone: string;
  /** @nullable */
  birthDate?: string | null;
  insuranceType: CreatePatientBodyInsuranceType;
  /** @nullable */
  insuranceName?: string | null;
  totalSessions: number;
  /** @nullable */
  notes?: string | null;
}

export type UpdatePatientBodyInsuranceType =
  (typeof UpdatePatientBodyInsuranceType)[keyof typeof UpdatePatientBodyInsuranceType];

export const UpdatePatientBodyInsuranceType = {
  particular: "particular",
  convenio: "convenio",
} as const;

export interface UpdatePatientBody {
  name?: string;
  phone?: string;
  /** @nullable */
  birthDate?: string | null;
  insuranceType?: UpdatePatientBodyInsuranceType;
  /** @nullable */
  insuranceName?: string | null;
  totalSessions?: number;
  remainingSessions?: number;
  /** @nullable */
  notes?: string | null;
}

export interface Therapist {
  id: number;
  name: string;
  specialty: string;
  phone: string;
  /** @nullable */
  availableHours?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTherapistBody {
  name: string;
  specialty: string;
  phone: string;
  /** @nullable */
  availableHours?: string | null;
}

export interface UpdateTherapistBody {
  name?: string;
  specialty?: string;
  phone?: string;
  /** @nullable */
  availableHours?: string | null;
}

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  agendado: "agendado",
  confirmado: "confirmado",
  presente: "presente",
  falta: "falta",
  cancelado: "cancelado",
  remarcado: "remarcado",
  encaixe: "encaixe",
} as const;

export interface Appointment {
  id: number;
  patientId: number;
  therapistId: number;
  date: string;
  time: string;
  status: AppointmentStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  originalAppointmentId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface AppointmentWithDetails {
  id: number;
  patientId: number;
  therapistId: number;
  date: string;
  time: string;
  status: AppointmentStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  originalAppointmentId?: number | null;
  createdAt: string;
  updatedAt: string;
  patientName: string;
  patientPhone: string;
  therapistName: string;
  therapistSpecialty: string;
}

export interface CreateAppointmentBody {
  patientId: number;
  therapistId: number;
  date: string;
  time: string;
  status: AppointmentStatus;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateAppointmentBody {
  patientId?: number;
  therapistId?: number;
  date?: string;
  time?: string;
  status?: AppointmentStatus;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateStatusBody {
  status: AppointmentStatus;
}

export interface RescheduleBody {
  date: string;
  time: string;
  /** @nullable */
  therapistId?: number | null;
}

export interface DashboardSummary {
  totalToday: number;
  totalAbsences: number;
  totalCompleted: number;
  totalScheduled: number;
  totalCancelled: number;
}

export interface DailyReportItem {
  date: string;
  total: number;
  completed: number;
  absent: number;
  cancelled: number;
}

export interface AbsenceReportItem {
  date: string;
  patientName: string;
  therapistName: string;
  time: string;
}

export interface SessionsReportItem {
  patientId: number;
  patientName: string;
  totalSessions: number;
  remainingSessions: number;
  usedSessions: number;
}

export type ListPatientsParams = {
  search?: string;
};

export type ListAppointmentsParams = {
  date?: string;
  weekStart?: string;
  therapistId?: number;
  status?: string;
  patientId?: number;
};

export type GetDailyReportParams = {
  startDate: string;
  endDate: string;
};

export type GetAbsenceReportParams = {
  startDate: string;
  endDate: string;
};