# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VitalFisio Clinic Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: weekStart
          in: query
          schema:
            type: string
        - name: therapistId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentWithDetails"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetails"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /appointments/{id}/status:
    patch:
      operationId: updateAppointmentStatus
      tags: [appointments]
      summary: Update appointment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /appointments/{id}/reschedule:
    post:
      operationId: rescheduleAppointment
      tags: [appointments]
      summary: Reschedule an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleBody"
      responses:
        "200":
          description: Rescheduled appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
      required:
        - id
        - username
        - name
        - role
...
    AppointmentStatus:
      type: string
      enum:
        - agendado
        - confirmado
        - presente
        - falta
        - cancelado
        - remarcado
        - encaixe
...
    AppointmentWithDetails:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        therapistId:
          type: integer
        date:
          type: string
        time:
          type: string
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        notes:
          type: ["string", "null"]
        originalAppointmentId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        patientName:
          type: string
        patientPhone:
          type: string
        therapistName:
          type: string
        therapistSpecialty:
          type: string
      required:
        - id
        - patientId
        - therapistId
        - date
        - time
        - status
        - createdAt
        - updatedAt
        - patientName
        - patientPhone
        - therapistName
        - therapistSpecialty
    CreateAppointmentBody:
      type: object
      properties:
        patientId:
          type: integer
        therapistId:
          type: integer
        date:
          type: string
        time:
          type: string
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        notes:
          type: ["string", "null"]
      required:
        - patientId
        - therapistId
        - date
        - time
        - status
    UpdateAppointmentBody:
      type: object
      properties:
        patientId:
          type: integer
        therapistId:
          type: integer
        date:
          type: string
        time:
          type: string
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        notes:
          type: ["string", "null"]
    UpdateStatusBody:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/AppointmentStatus"
      required:
        - status
    RescheduleBody:
      type: object
      properties:
        date:
          type: string
        time:
          type: string
        therapistId:
          type: ["integer", "null"]
      required:
        - date
        - time
...