# lib/api-zod/src/generated/types/appointmentStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VitalFisio Clinic Management API
 * OpenAPI spec version: 0.1.0
 */

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  agendado: "agendado",
  confirmado: "confirmado",
  presente: "presente",
  falta: "falta",
  cancelado: "cancelado",
  remarcado: "remarcado",
  encaixe: "encaixe",
} as const;