# lib/api-zod/src/generated/types/createAppointmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VitalFisio Clinic Management API
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentStatus } from "./appointmentStatus";

export interface CreateAppointmentBody {
  patientId: number;
  therapistId: number;
  date: string;
  time: string;
  status: AppointmentStatus;
  /** @nullable */
  notes?: string | null;
}